feat: cập nhật

+++ src/main/java/com/example/demo/repository/ProductRepository.java
@@ -1,22 +1,29 @@
 package com.example.demo.repository;
 
-import com.example.demo.entity.Product;
+import java.math.BigDecimal;
+import java.util.List;
+
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 import org.springframework.stereotype.Repository;
 
-import java.util.List;
+import com.example.demo.entity.Product;
 
 @Repository
 public interface ProductRepository extends JpaRepository<Product, Long> {
+    // Lọc theo Category chuẩn đét
     List<Product> findByCategoryId(Long categoryId);
+
     List<Product> findByNameContainingIgnoreCase(String name);
-@Query("SELECT p FROM Product p WHERE " +
+
+    @Query("SELECT p FROM Product p WHERE " +
            "LOWER(p.name) LIKE LOWER(CONCAT('%', :keyword, '%')) OR " +
            "LOWER(p.description) LIKE LOWER(CONCAT('%', :keyword, '%'))")
     List<Product> searchProducts(@Param("keyword") String keyword);
-@Query("SELECT p FROM Product p WHERE p.price BETWEEN :minPrice AND :maxPrice")
-    List<Product> findByPriceRange(@Param("minPrice") Double minPrice, 
-                                    @Param("maxPrice") Double maxPrice);
+
+    // Sửa Double thành BigDecimal để xử lý tiền lớn không giới hạn
+    @Query("SELECT p FROM Product p WHERE p.price BETWEEN :minPrice AND :maxPrice")
+    List<Product> findByPriceRange(@Param("minPrice") BigDecimal minPrice, 
+                                   @Param("maxPrice") BigDecimal maxPrice);
 }
